app/backend/requirements.txt
fastapi
uvicorn
pypdf
python-docx
google.generativeai
openai
sentence-transformers
faiss-cpu
numpy